Add a Spring Boot application to run functional tests.
This commit adds a Spring Boot application that invokes JUnit on start
up.  It then runs the tests in the application.  The overall idea is to
allow this application to act as a custom test harness that we can
configure to our own needs and use to run functional tests against a
running instance of Candlepin.