Return a non-zero exit code if tests fail.
The JUnitBootstrap class now publishes an event with the summary of the
test results.  We have a class that listens to this event and totals the
number of failures reported.  This class also implements the
ExitCodeGenerator class and when it is time to exit, it examines the
total number of failures and determines the code to use based on that.

This is a rather complex way of doing things.  We could just throw an
exception that also implements ExitCodeGenerator.  However, when Spring
receives an exception in the Spring.exit method, it prints the
stacktrace.  This results in an unsightly stacktrace at the end of the
test run.  The trace doesn't provide any useful information and pushes
actual useful information off the screen.  Therefore, I went with the
more complicated, but more attractive approach.